requirements.txt
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
ipython==8.15.0
jedi==0.18.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==24.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure_eval==0.2.2
pygments==2.15.1
setuptools==69.5.1
six==1.16.0
stack_data==0.2.0
traitlets==5.7.1
typing_extensions==4.11.0
wcwidth==0.2.5
wheel==0.43.0
Pillow==10.3.0
split-folders==0.5.1
torch==2.3.0
PyYAML==6.0.1
opencv-python==4.9.0.80
transformers==4.41.1
ultralytics==8.2.20
pyrealsense2==2.55.1.6486
imgaug==0.4.0
albumentations==1.4.7
logging
argparse